Add framework for all routes and verified basic GET integrity with connection sql database with MYSQL2 npm package
