+++ routes/api/routes_Department.js
@@ -3,20 +3,27 @@
 
 const router = require('express').Router();
 
+//-- SQL database connection with mysql2
+const db = require('../../config/connection')
+
 //------------------------------------------------------------------------------
 //-- ROUTES
 
 router.get('/', async (req, res) => {
-    try {
-      // find all library cards and perform a JOIN to include all associated Readers
-      const libraryCardData = await LibraryCard.findAll({
-        include: [{ model: Reader }]
-      });
-      res.status(200).json(libraryCardData);
-    } catch (err) {
-      res.status(500).json(err);
-    }
-  });
+  try {
+      // Query database for ALL employees with all columns
+    db.query('SELECT * FROM department', function (err, results) {
+      if (err) {
+        res.status(500).json(err);
+      }
+      res.status(200).json(results);
+    });
+  }
+  catch (err) {
+    res.status(500).json(err);
+  }
+});
+
 
 
 //------------------------------------------------------------------------------
